adding prodcuts to cart

+++ app/src/main/java/com/cabify/ui/screen/CartScreen.kt
@@ -1,65 +1,49 @@
 package com.cabify.ui.screen
 
-import android.graphics.drawable.shapes.OvalShape
+import android.widget.Toast
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.ClickableText
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonDefaults
 import androidx.compose.material.Card
-import androidx.compose.material.ExtendedFloatingActionButton
 import androidx.compose.material.FabPosition
-import androidx.compose.material.FloatingActionButton
 import androidx.compose.material.Icon
 import androidx.compose.material.IconButton
-import androidx.compose.material.IconToggleButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Scaffold
 import androidx.compose.material.Text
 import androidx.compose.material.TopAppBar
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Clear
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.material.icons.filled.FavoriteBorder
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
 import coil.compose.rememberAsyncImagePainter
 import com.cabify.R
 import com.cabify.common.toEuro
@@ -71,7 +55,6 @@ import com.cabify.ui.theme.medium
 import com.cabify.ui.theme.normal
 import com.cabify.ui.theme.regular
 import com.cabify.ui.theme.small
-import com.cabify.ui.theme.xlarge
 import com.cabify.ui.theme.xsmall
 import com.cabify.ui.theme.xxlarge
 import com.cabify.viewmodel.ProductViewModel
@@ -81,11 +64,8 @@ fun CartScreen(
     viewModel: ProductViewModel,
     onBack: () -> Unit
 ) {
-
-    val selectedProducts by viewModel.selectedProductsState.collectAsState()
-    var fabHeight by remember {
-        mutableStateOf(0)
-    }
+    val context = LocalContext.current
+    val productState by viewModel.productState.collectAsState()
 
     Scaffold(
         topBar = {
@@ -98,6 +78,7 @@ fun CartScreen(
         },
         floatingActionButton = {
             Button(
+                enabled = viewModel.countAddedProducts.value != 0,
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(horizontal = normal),
@@ -105,10 +86,12 @@ fun CartScreen(
                 colors = ButtonDefaults.buttonColors(backgroundColor = Purple700),
                 shape = RoundedCornerShape(28.dp)
             ) {
-                Column(modifier = Modifier
-                    .fillMaxWidth(),
-                verticalArrangement = Arrangement.SpaceBetween,
-                horizontalAlignment = Alignment.CenterHorizontally) {
+                Column(
+                    modifier = Modifier
+                        .fillMaxWidth(),
+                    verticalArrangement = Arrangement.SpaceBetween,
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
                     Text(
                         text = stringResource(id = R.string.total_without_discount, "12500€"),
                         style = MaterialTheme.typography.caption.copy(
@@ -133,31 +116,67 @@ fun CartScreen(
         floatingActionButtonPosition = FabPosition.Center
     ) { paddingValues ->
         Column(Modifier.padding(paddingValues)) {
-            LazyColumn(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(bottom = xxlarge),
-                state = rememberLazyListState()
-            ) {
-                itemsIndexed(selectedProducts) { _, product ->
-                    ItemAddedProduct(
-                        product = product,
-                        onAdd = {
-                            viewModel.addProduct(it)
-                        }, onDelete = {
-                            viewModel.deleteProduct(it)
-                        })
+            if (viewModel.countAddedProducts.value != 0) {
+                LazyColumn(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(bottom = xxlarge),
+                    state = rememberLazyListState()
+                ) {
+                    itemsIndexed(productState.data) { _, product ->
+                        if (product.itemAdded.value != 0) {
+                            ItemAddedProduct(
+                                product = product,
+                                onAdd = {
+                                    viewModel.addProduct(it)
+                                }, onDelete = {
+                                    viewModel.deleteProduct(it)
+                                })
+                        }
+                    }
+                }
+            } else {
+                Column(
+                    modifier = Modifier.fillMaxSize(),
+                    verticalArrangement = Arrangement.Center,
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Text(
+                        text = stringResource(id = R.string.empty_cart),
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(normal),
+                        maxLines = 2,
+                        overflow = TextOverflow.Ellipsis,
+                        style = MaterialTheme.typography.h5.copy(
+                            color = Color.Black,
+                            fontWeight = FontWeight.Light,
+                            textAlign = TextAlign.Center
+                        )
+                    )
+
+                    ClickableText(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .padding(normal),
+                        text = buildAnnotatedString { append(stringResource(id = R.string.see_products)) },
+                        maxLines = 2,
+                        overflow = TextOverflow.Ellipsis,
+                        style = MaterialTheme.typography.h6.copy(
+                            color = Purple700,
+                            fontWeight = FontWeight.Bold,
+                            textAlign = TextAlign.Center
+                        ),
+                        onClick = { onBack() }
+                    )
                 }
             }
+        }
 
-            FloatingActionButton(
-                modifier = Modifier
-                    .padding(all = 16.dp),
-                onClick = {
-
-                }) {
-                Icon(imageVector = Icons.Filled.Add, contentDescription = "Add")
-            }
+        if (viewModel.showMessage.value) {
+            Toast.makeText(context, stringResource(id = R.string.no_stock), Toast.LENGTH_LONG)
+                .show()
+            viewModel.showMessage.value = false
         }
     }
 }
@@ -228,7 +247,10 @@ fun ItemAddedProduct(
                     )
 
                     Text(
-                        text = stringResource(id = R.string.stock, product.stock),
+                        text = stringResource(
+                            id = R.string.stock,
+                            product.stock - product.itemAdded.value
+                        ),
                         modifier = Modifier
                             .padding(start = normal, top = medium, end = normal)
                             .align(Alignment.End),
@@ -270,7 +292,7 @@ fun ItemAddedProduct(
                         }
 
                         Text(
-                            text = "1",
+                            text = product.itemAdded.value.toString(),
                             style = MaterialTheme.typography.subtitle2.copy(
                                 color = Color.Black,
                                 fontWeight = FontWeight.SemiBold
@@ -294,7 +316,7 @@ fun ItemAddedProduct(
                 Text(
                     text = stringResource(id = R.string.discount, "4500€"),
                     style = MaterialTheme.typography.caption.copy(
-                        color = Color.Red,
+                        color = Purple700,
                         fontWeight = FontWeight.Bold
                     ),
                     modifier = Modifier.padding(horizontal = normal)
